requirements.txt
# Core Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Visualization
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine Learning & Neural Networks
torch>=1.9.0
xgboost>=1.5.0
lightgbm>=3.3.0

# AutoML
pycaret

# Time Series
prophet>=1.0.0

# NLP & Text Processing
transformers>=4.15.0
sentence-transformers>=2.0.0
nltk>=3.6.0

# Data Profiling & Analysis
ydata-profiling>=4.0.0
shap>=0.40.0

# Streamlit & Web Framework
streamlit>=1.22.0
streamlit-aggrid>=0.3.0
streamlit-option-menu>=0.3.0

# MongoDB
pymongo>=4.0.0
dnspython>=2.0.0

# Database & APIs
sqlalchemy>=1.4.0
requests>=2.25.0

# Utilities
python-dotenv>=0.19.0
joblib>=1.1.0
tqdm>=4.62.0